moved navbar container out of appconainter and styled the navbar
