+++ browser/Containers/CSSEditor.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 const CSSEditor = props => {
   return (
-    <div className="container-fluid" style={props.style}>
+    <div className="container-fluid">
       <div className="row">
         <div className="col-md-12">
           <div id="css-firepad-container"></div>
