+++ browser/Containers/HTMLEditor.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 const HTMLEditor = props => {
   return (
-    <div className="container-fluid" style={props.style} >
+    <div className="container-fluid">
       <div className="row">
         <div className="col-md-12">
           <div id="html-firepad-container"></div>
